factor: Split the CLI and I/O code off the factoring logic
