+++ src/uu/factor/Cargo.toml
@@ -11,9 +11,6 @@ keywords = ["coreutils", "uutils", "cross-platform", "cli", "utility"]
 categories = ["command-line-utilities"]
 edition = "2018"
 
-[lib]
-path = "src/factor.rs"
-
 [dependencies]
 rand = "0.5"
 uucore = { version="0.0.4", package="uucore", git="https://github.com/uutils/uucore.git", branch="canary" }
@@ -22,3 +19,6 @@ uucore_procs = { version="0.0.4", package="uucore_procs", git="https://github.co
 [[bin]]
 name = "factor"
 path = "src/main.rs"
+
+[lib]
+path = "src/cli.rs"
